refactor: move kmx_u16 to common and rename to km_u16
