+++ common/include/km_u16.h
