+++ developer/src/kmcmplib/src/meson.build
@@ -78,13 +78,13 @@ lib = library('kmcmplib',
   'cp1252.cpp',
   'DeprecationChecks.cpp',
   'Edition.cpp',
-  'kmx_u16.cpp',
   'NamedCodeConstants.cpp',
   'UnreachableRules.cpp',
   'uset-api.cpp',
   'versioning.cpp',
   'virtualcharkeys.cpp',
   'xstring.cpp',
+  '../../../../common/cpp/km_u16.cpp',
   '../../../../common/cpp/utfcodec.cpp',
   '../../../../common/windows/cpp/src/ConvertUTF.c',
   '../../../../common/windows/cpp/src/crc32.cpp',
