+++ developer/src/kmcmplib/src/pch.h
@@ -4,7 +4,7 @@
 
 #define USE_CHAR16_T
 #include <kmx_file.h>
-#include "kmx_u16.h"
+#include <km_u16.h>
 #include <kmcompx.h>
 
 #include "../../../../common/windows/cpp/include/crc32.h"
